phdf5write: create a file and a simple dataset. Writing frames not yet fully working.

+++ phdf5writeApp/src/dimension.h
@@ -16,7 +16,7 @@
 
 class NDArray; // forward declaration
 
-typedef unsigned long int dimsize_t;
+typedef unsigned long long dimsize_t;
 typedef std::vector<dimsize_t> vec_ds_t;
 
 // Describe an N-dimension blob.
